Split signal module in manager, impl and test
